Rename diskmaker to devicefinder
