+++ internal/devicefinder/event_reporter.go
@@ -1,4 +1,4 @@
-package diskmaker
+package devicefinder
 
 import (
 	"fmt"
